refactor: change to use quote! and proc_macro2
proc_macro2 allows the implementation of separate modules.
