+++ packages/derive/Cargo.toml
@@ -17,6 +17,7 @@ default = []
 [dependencies]
 syn = { version = "1.0", features = ["full"] }
 quote = "1.0"
+proc-macro2 = "1.0"
 proc-macro-error = { version = "1", default-features = false }
 
 [dev-dependencies]
